refactor: rename useScroller to useYScroller
